+++ margo/src/protocols/mod.rs
@@ -6,6 +6,7 @@ pub mod gamma_control;
 pub mod generated;
 pub mod layer_shell;
 pub mod output_management;
+pub mod output_power;
 pub mod screencopy;
 pub mod virtual_pointer;
 pub mod wlr_foreign_toplevel;